+++ packages/root/types/types/index.d.ts
@@ -1,6 +1,6 @@
 export { HistoryOptions, HistoryConstructor, History } from "./hickory";
 export { LocationComponents, PartialLocation, SessionLocation, Hrefable, Key } from "./location";
 export { KeyFns } from "./key_generator";
-export { LocationUtilOptions, QueryFunctions, VerifyPathname, LocationUtils } from "./location_utils";
+export { LocationUtilOptions, QueryFunctions, LocationUtils } from "./location_utils";
 export { NavigationInfo, ConfirmationFunction, ConfirmationMethods, BlockingHistory } from "./navigation_confirmation";
 export { ToArgument, Action, NavType, FinishNavigation, CancelNavigation, PendingNavigation, ResponseHandler, Preparer, FinishCancel, NavigateArgs, NavigateHelpers } from "./navigate";